Consolidated local node "gdrive" and "oauthclient" modules into a single "gdrive" module. Revised corresponding code.
