--- lib/oauthclient/package.json
@@ -1,10 +1,0 @@
-{
-  "name": "oauthclient",
-  "version": "1.0.0",
-  "description": "Google API authenication client",
-  "main": "oauthclient.js",
-  "author": "Chris Pels",
-  "dependencies": {
-    "googleapis":"24.0.0"
-  }
-}
