+++ package.json
@@ -12,22 +12,19 @@
   "author": "Brad Noble @bradnoble",
   "license": "Apache-2.0",
   "dependencies": {
-    "body-parser": "^1.17.2",
-    "cfenv": "^1.0.4",
-    "cloudant": "^1.8.0",
-    "dotenv": "^4.0.0",
-    "express": "^4.15.3",
-    "express-session": "^1.15.6",
+    "body-parser": "1.17.2",
+    "cfenv": "1.0.4",
+    "cloudant": "1.8.0",
+    "dotenv": "4.0.0",
+    "express": "4.15.3",
+    "express-session": "1.15.6",
     "gdrive": "file:lib/gdrive",
-    "googleapis": "^25.0.0",
     "http": "0.0.0",
-    "json2csv": "^3.10.0",
-    "materialize-css": "^1.0.0-beta",
-    "memory-streams": "^0.1.3",
-    "oauthclient": "file:lib/oauthclient",
-    "passport": "^0.3.2",
-    "passport-http": "^0.3.0",
-    "passport-local": "^1.0.0"
+    "json2csv": "3.10.0",
+    "memory-streams": "0.1.3",
+    "passport": "0.3.2",
+    "passport-http": "0.3.0",
+    "passport-local": "1.0.0"
   },
   "repository": {
     "type": "git",
